[feat] Re-write mongodb connector to allow more flexible configs
Main goal: allow connecting to mongodb using SSL (the old connector
didn't allow this kind of thing).

Other features:
- folder for automated tests;
- load of certificate file content into connetion setting (for mongo
only);

Note: we tried to upgrade node-mongodb driver, but the new version (2.2)
was failing on a lot of benchmark tests, while the current version (2.0)
was stable and didn't have any failure during some runnings of benchmark
scripts. The new version uses a much simpler structure of the
settings.extra, so it would be great if we could upgrade it. To make
this change easier, the code (and tests) are ready for both versions, so
whenever v.2.2 is stable we might change it here just by removing some
lines of code.