lint: Fix ESLint guard-for-in errors
`for...in` iterates over all enumerable properties including inherited
properties. This is almost never what is desired. Tweak the code to
iterate via `for...of` or `Array.prototype.map()`.

This is a re-do of commit 2d1024fe32d6fc00f6bf51914083eb98356d7650
which was reverted in commit a52ee0d2284036e5e85db152965022321f67cb6a
due to bugs.