cache: Fix overly aggressive cache entry deletion in `gc()`
For some mysterious reason `gc()` was deleting entries until the cache
was half full. Now it only deletes entries until the cache is no
longer overfull.

Maybe the rationale behind deleting extra is to amortize the cost of
`gc()`. However, if there are more than `settings.cache` dirty entries
then `gc()` will run anyway so it doesn't help when the system is
heavily loaded (and it could hurt by reducing the effective size of
the cache).