tests: Fix error handling in `promisifyBufferedFn()`
Before this change, an error passed to the buffered callback would
result in an unhandled Promise rejection if the caller was only
awaiting the written callback. Similarly, an error passed to the
written callback would result in an unhandled Promise rejection if the
caller was only awaiting the buffered callback.