cache: Always mark the cache entry as most recently used in `set()`
This change causes more calls to `evictOld()`, which could reduce
performance. If this is a problem, we can change the logic to only run
`evictOld()` if a new entry is being added, or we can speed up
`evictOld()` by maintaining a separate list of evictable
entries (which we should investigate anyway).