tests: Increase `setMax` for `postgres`
It's unclear why this is necessary as none of the recent changes
should have negatively impacted performance. There's enough variance
in the benchmarks that it could have been bad luck. Because we don't
have good benchmarking, these limits are only good as a smoke test, so
it should be OK to increase the limit.