mysql: Let the `mysql` package build the SQL strings
This simplifies the code, and might be more efficient than the
previous algorithm. (The previous algorithm appears to be O(N^2) due
to incremental string building, but when I ran benchmarks the time
didn't appear to grow superlinearly. Apparently the V8 engine has lots
of clever optimizations for string concatenation.)